requirements.txt
streamlit>=1.32.0
boto3>=1.34.0
python-dotenv>=1.0.0
pandas>=2.0.0
Pillow>=10.0.0
python-magic>=0.4.27
requests>=2.31.0
typing-extensions>=4.5.0
gradio>=4.0.0
moviepy>=1.0.3
